Changed Expression -> added inner logic and inner ValueHolder class

+++ src/main/java/com/sysgears/simplecalculator/computer/Operators.java
@@ -3,25 +3,25 @@ package com.sysgears.simplecalculator.computer;
 public enum Operators {
     PLUS('+', 100) {
         @Override
-        public Integer evaluate(Expression expression) {
+        public Double evaluate(Expression expression) {
             return expression.getFirstValue() + expression.getSecondValue();
         }
     },
     MINUS('-', 100) {
         @Override
-        public Integer evaluate(Expression expression) {
+        public Double evaluate(Expression expression) {
             return expression.getFirstValue() - expression.getSecondValue();
         }
     },
     MULTIPLAY('*', 1000) {
         @Override
-        public Integer evaluate(Expression expression) {
+        public Double evaluate(Expression expression) {
             return expression.getFirstValue() * expression.getSecondValue();
         }
     },
     DIVIDE('/', 1000) {
         @Override
-        public Integer evaluate(Expression expression) {
+        public Double evaluate(Expression expression) {
             return expression.getFirstValue() / expression.getSecondValue();
         }
     };
@@ -38,5 +38,5 @@ public enum Operators {
         return priority;
     }
 
-    public abstract Integer evaluate(Expression expression);
+    public abstract Double evaluate(Expression expression);
 }
